+++ src/TreeNode.js
@@ -1341,7 +1341,7 @@ class TreeNode extends ImmutableNode {
   }
 
   static getVersion() {
-    return "9.0.0"
+    return "9.1.0"
   }
 
   static getPathWithoutFileName(path) {